兼容性方法添加@SuppressWarnings("deprecation")

+++ src/main/java/cc/azuramc/bedwars/compat/wrapper/MaterialWrapper.java
@@ -353,9 +353,9 @@ public class MaterialWrapper {
                     .getItem();
         }
     }
-    
+
     public static byte getWoolData(DyeColor dyeColor) {
-        return (byte) dyeColor.getWoolData();
+        return dyeColor.getWoolData();
     }
     
     public static Material BED() {
